requirements.txt
pycurl
docker-compose
github_webhook
sqlalchemy
sqlalchemy-datatables